+++ yt/visualization/tests/test_callbacks.py
@@ -150,7 +150,7 @@ def test_line_callback():
         # Now we'll check a few additional minor things
         p = SlicePlot(ds, "x", "density")
         p.annotate_line(
-            [0.1, 0.1], [0.5, 0.5], coord_system="axis", plot_args={"color": "red"}
+            [0.1, 0.1], [0.5, 0.5], coord_system="axis", mpl_kwargs={"color": "red"}
         )
         p.save(prefix)
 
@@ -186,7 +186,7 @@ def test_ray_callback():
         # Now we'll check a few additional minor things
         p = SlicePlot(ds, "x", "density")
         p.annotate_ray(oray)
-        p.annotate_ray(ray, plot_args={"color": "red"})
+        p.annotate_ray(ray, mpl_kwargs={"color": "red"})
         p.save(prefix)
 
     with _cleanup_fname() as prefix:
@@ -578,7 +578,7 @@ def test_contour_callback():
             factor=8,
             take_log=False,
             clim=(0.4, 0.6),
-            plot_args={"linewidths": 2.0},
+            mpl_kwargs={"linewidths": 2.0},
             label=True,
             text_args={"fontsize": "x-large"},
         )
@@ -592,7 +592,7 @@ def test_contour_callback():
             factor=8,
             take_log=False,
             clim=(0.4, 0.6),
-            plot_args={"linewidths": 2.0},
+            mpl_kwargs={"linewidths": 2.0},
             label=True,
             text_args={"fontsize": "x-large"},
             data_source=s2,
@@ -609,7 +609,7 @@ def test_contour_callback():
             take_log=False,
             clim=(1.0e-1, 1.0e1),
             label=True,
-            plot_args={"colors": ("c", "w"), "linewidths": 1},
+            mpl_kwargs={"colors": ("c", "w"), "linewidths": 1},
             text_args={"fmt": "%1.1f"},
         )
         assert_fname(slc.save(prefix)[0])
@@ -623,7 +623,7 @@ def test_contour_callback():
             factor=8,
             take_log=False,
             clim=(0.4, 0.6),
-            plot_args={"linewidths": 2.0},
+            mpl_kwargs={"linewidths": 2.0},
             label=True,
             text_args={"fontsize": "x-large"},
         )
@@ -720,13 +720,13 @@ def test_mesh_lines_callback():
         ds = fake_hexahedral_ds()
         for field in ds.field_list:
             sl = SlicePlot(ds, 1, field)
-            sl.annotate_mesh_lines(plot_args={"color": "black"})
+            sl.annotate_mesh_lines(mpl_kwargs={"color": "black"})
             assert_fname(sl.save(prefix)[0])
 
         ds = fake_tetrahedral_ds()
         for field in ds.field_list:
             sl = SlicePlot(ds, 1, field)
-            sl.annotate_mesh_lines(plot_args={"color": "black"})
+            sl.annotate_mesh_lines(mpl_kwargs={"color": "black"})
             assert_fname(sl.save(prefix)[0])
 
 
@@ -769,7 +769,7 @@ def test_streamline_callback():
                 "velocity_y",
                 field_color="magvel",
                 display_threshold=0.5,
-                plot_args={
+                mpl_kwargs={
                     "cmap": ytcfg.get("yt", "default_colormap"),
                     "arrowstyle": "->",
                 },